@@ -20,31 +20,31 @@
     "start": "next start"
   },
   "dependencies": {
-    "@react-three/drei": "^9.56.5",
-    "@react-three/fiber": "^8.10.1",
+    "@react-three/drei": "^9.56.27",
+    "@react-three/fiber": "^8.11.5",
     "glsl-random": "^0.0.5",
-    "next": "^13.1.6",
+    "next": "^13.2.1",
     "next-pwa": "^5.6.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "three": "^0.149.0",
-    "three-stdlib": "^2.21.6"
+    "three": "^0.150.0",
+    "three-stdlib": "^2.21.8",
+    "tunnel-rat": "^0.1.0"
   },
   "devDependencies": {
     "@next/bundle-analyzer": "^13.1.6",
     "autoprefixer": "^10.4.12",
-    "eslint": "^8.33.0",
-    "eslint-config-next": "^13.1.6",
+    "eslint": "^8.34.0",
+    "eslint-config-next": "^13.2.1",
     "eslint-config-prettier": "^8.6.0",
     "eslint-plugin-tailwindcss": "^3.9.0",
-
     "file-loader": "^6.2.0",
     "glslify": "^7.1.1",
     "glslify-loader": "^2.0.0",
     "postcss": "^8.4.21",
-    "prettier": "^2.8.3",
+    "prettier": "^2.8.4",
     "raw-loader": "^4.0.2",
-    "tailwindcss": "^3.2.4",
+    "tailwindcss": "^3.2.7",
     "url-loader": "^4.1.1"
   }
 }
