@@ -1,5 +1,8 @@
+'use client'
+
 import { Canvas } from '@react-three/fiber'
 import { OrbitControls, Preload } from '@react-three/drei'
+import { r3f } from '@/helpers/global'
 
 export default function Scene({ children, ...props }) {
   // Everything defined in here will persist between route changes, only children are swapped
@@ -7,7 +10,7 @@ export default function Scene({ children, ...props }) {
     <Canvas {...props}>
       <directionalLight intensity={0.75} />
       <ambientLight intensity={0.75} />
-      {children}
+      <r3f.Out />
       <Preload all />
       <OrbitControls />
     </Canvas>
